Move Filter code to dispatcher
