+++ data-plane/core/pom.xml
@@ -99,10 +99,6 @@
       <artifactId>cloudevents-http-vertx</artifactId>
     </dependency>
     <dependency>
-      <groupId>io.cloudevents</groupId>
-      <artifactId>cloudevents-sql</artifactId>
-    </dependency>
-    <dependency>
       <groupId>io.vertx</groupId>
       <artifactId>vertx-micrometer-metrics</artifactId>
       <exclusions>
