+++ data-plane/dispatcher/src/main/java/dev/knative/eventing/kafka/broker/dispatcher/impl/filter/AllFilter.java
@@ -13,18 +13,27 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.knative.eventing.kafka.broker.core.filter.impl;
+package dev.knative.eventing.kafka.broker.dispatcher.impl.filter;
 
+import dev.knative.eventing.kafka.broker.dispatcher.Filter;
 import io.cloudevents.CloudEvent;
+import java.util.Set;
 
-public class ExactFilter extends BaseStringFilter {
+public class AllFilter implements Filter {
 
-  public ExactFilter(String attribute, String expectedValue) {
-    super(attribute, expectedValue);
+  private final Set<Filter> filters;
+
+  public AllFilter(Set<Filter> filters) {
+    this.filters = filters;
   }
 
   @Override
   public boolean test(CloudEvent cloudEvent) {
-    return expectedValue.equals(this.extractor.apply(cloudEvent));
+    for (Filter filter : filters) {
+      if (!filter.test(cloudEvent)) {
+        return false;
+      }
+    }
+    return true;
   }
 }
