+++ data-plane/dispatcher/src/main/java/dev/knative/eventing/kafka/broker/dispatcher/impl/filter/ExactFilter.java
@@ -13,19 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.knative.eventing.kafka.broker.core.filter.impl;
+package dev.knative.eventing.kafka.broker.dispatcher.impl.filter;
 
 import io.cloudevents.CloudEvent;
 
-public class PrefixFilter extends BaseStringFilter {
+public class ExactFilter extends BaseStringFilter {
 
-  public PrefixFilter(String attribute, String expectedValue) {
+  public ExactFilter(String attribute, String expectedValue) {
     super(attribute, expectedValue);
   }
 
   @Override
   public boolean test(CloudEvent cloudEvent) {
-    String value = this.extractor.apply(cloudEvent);
-    return value != null && value.startsWith(this.expectedValue);
+    return expectedValue.equals(this.extractor.apply(cloudEvent));
   }
 }
