+++ data-plane/dispatcher/src/test/java/dev/knative/eventing/kafka/broker/dispatcher/impl/filter/SqlFilterTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.knative.eventing.kafka.broker.core.filter.impl;
+package dev.knative.eventing.kafka.broker.dispatcher.impl.filter;
 
 import io.cloudevents.CloudEvent;
 import io.cloudevents.core.builder.CloudEventBuilder;
@@ -27,7 +27,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class PrefixFilterTest {
+public class SqlFilterTest {
 
   final static CloudEvent event = CloudEventBuilder.v1()
     .withId("123-42")
@@ -45,18 +45,20 @@ public class PrefixFilterTest {
 
   @ParameterizedTest
   @MethodSource(value = {"testCases"})
-  public void match(CloudEvent event, String key, String value, boolean shouldMatch) {
-    var filter = new PrefixFilter(key, value);
+  public void match(CloudEvent event, String expression, boolean shouldMatch) {
+    var filter = new SqlFilter(expression);
     assertThat(filter.test(event))
       .isEqualTo(shouldMatch);
   }
 
   static Stream<Arguments> testCases() {
     return Stream.of(
-      Arguments.of(event, "id", "123", true),
-      Arguments.of(event, "id", "124", false),
-      Arguments.of(event, "source", "/api", true),
-      Arguments.of(event, "source", "/news", false)
+      Arguments.of(event, "'TRUE'", true),
+      Arguments.of(event, "'FALSE'", false),
+      Arguments.of(event, "0", false),
+      Arguments.of(event, "1", false),
+      Arguments.of(event, "id LIKE '123%'", true),
+      Arguments.of(event, "NOT(id LIKE '123%')", false)
     );
   }
