+++ data-plane/dispatcher/src/test/java/dev/knative/eventing/kafka/broker/dispatcher/impl/filter/SuffixFilterTest.java
@@ -13,16 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.knative.eventing.kafka.broker.core.filter.impl;
+package dev.knative.eventing.kafka.broker.dispatcher.impl.filter;
 
-import dev.knative.eventing.kafka.broker.core.filter.Filter;
 import io.cloudevents.CloudEvent;
 import io.cloudevents.core.builder.CloudEventBuilder;
 import java.net.URI;
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
-import java.util.Collections;
-import java.util.Set;
 import java.util.stream.Stream;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
@@ -30,7 +27,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class AnyFilterTest {
+public class SuffixFilterTest {
 
   final static CloudEvent event = CloudEventBuilder.v1()
     .withId("123-42")
@@ -48,19 +45,18 @@ public class AnyFilterTest {
 
   @ParameterizedTest
   @MethodSource(value = {"testCases"})
-  public void match(CloudEvent event, Filter filter, boolean shouldMatch) {
+  public void match(CloudEvent event, String key, String value, boolean shouldMatch) {
+    var filter = new SuffixFilter(key, value);
     assertThat(filter.test(event))
       .isEqualTo(shouldMatch);
   }
 
   static Stream<Arguments> testCases() {
     return Stream.of(
-      Arguments.of(event, new AnyFilter(Set.of(new ExactFilter("id", "123-42"))), true),
-      Arguments.of(event, new AnyFilter(Set.of(new ExactFilter("id", "123-42"), new ExactFilter("source", "/api/some-source"))), true),
-      Arguments.of(event, new AnyFilter(Set.of(new ExactFilter("id", "123"), new ExactFilter("source", "/api/some-source"))), true),
-      Arguments.of(event, new AnyFilter(Set.of(new ExactFilter("id", "123-42"), new ExactFilter("source", "/api/something-else"))), true),
-      Arguments.of(event, new AnyFilter(Set.of(new ExactFilter("id", "123"), new ExactFilter("source", "/api/something-else"))), false),
-      Arguments.of(event, new AnyFilter(Collections.emptySet()), false)
+      Arguments.of(event, "id", "42", true),
+      Arguments.of(event, "id", "43", false),
+      Arguments.of(event, "source", "source", true),
+      Arguments.of(event, "source", "sink", false)
     );
   }
